+++ path/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add a "parse" filter to get path components
+
 ## [2.2.0] - 2026-01-02
 
 ### Added